requirements.txt
flask==3.0.0
python-dotenv==1.0.0
python-dateutil==2.8.2
gunicorn==21.2.0
requests==2.31.0
praw==7.8.1 